+++ packages/renderer/src/test/derive-codec-from-file-extension.test.ts
